+++ CHANGES.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added `Grove.Buzzer`.
 
 ## 0.1.0 - 2016-08-15
 ### Added